# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Event:
    properties:
      additional_info:
        type: string
      address:
        type: string
      capacity:
        type: integer
      category:
        type: string
      checkin_token:
        type: string
      city:
        type: string
      created_at:
        type: string
      date_end:
        type: string
      date_start:
        type: string
      description:
        type: string
      format:
        $ref: '#/definitions/domain.EventFormat'
      id:
        type: integer
      is_free:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      organizer:
        $ref: '#/definitions/domain.User'
      organizer_contact:
        type: string
      organizer_id:
        type: integer
      poster_url:
        type: string
      price:
        type: number
      reg_deadline:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  domain.EventFormat:
    enum:
    - online
    - offline
    - hybrid
    type: string
    x-enum-varnames:
    - FormatOnline
    - FormatOffline
    - FormatHybrid
...
  domain.User:
    properties:
      approved:
        type: boolean
      avatar_url:
        type: string
      bio:
        type: string
      blocked:
        type: boolean
      city:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      grade:
        type: integer
      id:
        type: integer
      interests:
        items:
          type: string
        type: array
      phone:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
      school:
        type: string
      updated_at:
        type: string
      visible_to_organizers:
        type: boolean
      visible_to_school:
        type: boolean
    type: object
  domain.UserRole:
    enum:
    - student
    - organizer
    - admin
    type: string
    x-enum-varnames:
    - RoleStudent
    - RoleOrganizer
    - RoleAdmin
...
  dto.CreateEventRequest:
    properties:
      additional_info:
        type: string
      address:
        type: string
      capacity:
        type: integer
      category:
        type: string
      city:
        type: string
      date_end:
        type: string
      date_start:
        type: string
      description:
        type: string
      format:
        $ref: '#/definitions/domain.EventFormat'
      is_free:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      organizer_contact:
        type: string
      price:
        type: number
      reg_deadline:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - date_start
    - title
    type: object
  dto.EventDetailResponse:
    properties:
      additional_info:
        type: string
      address:
        type: string
      capacity:
        type: integer
      category:
        type: string
      checkin_token:
        type: string
      city:
        type: string
      created_at:
        type: string
      date_end:
        type: string
      date_start:
        type: string
      description:
        type: string
      format:
        $ref: '#/definitions/domain.EventFormat'
      free_seats:
        type: integer
      id:
        type: integer
      is_favorite:
        type: boolean
      is_free:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      organizer:
        $ref: '#/definitions/domain.User'
      organizer_contact:
        type: string
      organizer_id:
        type: integer
      poster_url:
        type: string
      price:
        type: number
      reg_deadline:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  dto.EventListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Event'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  dto.UpdateEventRequest:
    properties:
      additional_info:
        type: string
      address:
        type: string
      capacity:
        type: integer
      category:
        type: string
      city:
        type: string
      date_end:
        type: string
      date_start:
        type: string
      description:
        type: string
      format:
        $ref: '#/definitions/domain.EventFormat'
      is_free:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      organizer_contact:
        type: string
      poster_url:
        type: string
      price:
        type: number
      reg_deadline:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  response.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.ErrorBody:
    properties:
      error:
        $ref: '#/definitions/response.APIError'
    type: object
host: 192.168.0.103:8080
info:
  contact:
    email: support@eventapp.local
    name: EventApp Team
  description: Platform for school students and IT/robotics event organizers.
  license:
    name: MIT
  termsOfService: http://localhost:8080
  title: EventApp API
  version: "1.2"
paths:
...
  /api/events:
    get:
      description: Returns a paginated, filterable list of events. Supports advanced
        filters.
      parameters:
      - description: Filter by city
        in: query
        name: city
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by format
        in: query
        name: format
        type: string
      - description: Search title/description
        in: query
        name: search
        type: string
      - description: Comma-separated tags
        in: query
        name: tags
        type: string
      - description: true or false
        in: query
        name: is_free
        type: string
      - description: RFC3339 date
        in: query
        name: date_from
        type: string
      - description: RFC3339 date
        in: query
        name: date_to
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventListResponse'
      security:
      - BearerAuth: []
      summary: List events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Only approved organizers may create events. A unique checkin_token
        is generated automatically.
      parameters:
      - description: Event payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create a new event
      tags:
      - events
  /api/events/{id}:
    delete:
      description: Only the approved organizer who created the event can delete it.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - events
    get:
      description: Returns event details including free seats and favorite status
        for the caller.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get event by ID
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Only the approved organizer who created the event can update it.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Event'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update event
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"